Issue #419 - refactor Runner.print_issue() to utils.issue_string_tuple()
