+++ awslimitchecker/tests/test_runner.py
@@ -869,11 +869,12 @@ class TestCheckThresholds(RunnerTester):
         }
         mock_checker.get_limits.return_value = {}
 
-        def se_print(cls, s, l, c, w):
+        def se_print(s, l, c, w, colorize=True):
             return ('{s}/{l}'.format(s=s, l=l.name), '')
 
         self.cls.checker = mock_checker
-        with patch('awslimitchecker.runner.Runner.print_issue',
+        self.cls.colorize = False
+        with patch('%s.issue_string_tuple' % pb,
                    autospec=True) as mock_print:
             mock_print.side_effect = se_print
             with patch('awslimitchecker.runner.dict2cols') as mock_d2c:
@@ -883,10 +884,19 @@ class TestCheckThresholds(RunnerTester):
             call.check_thresholds(use_ta=True, service=None)
         ]
         assert mock_print.mock_calls == [
-            call(self.cls, 'svc1', mock_limit1, [mock_c1], [mock_w1]),
-            call(self.cls, 'svc1', mock_limit2, [], [mock_w2]),
-            call(self.cls, 'svc2', mock_limit3, [], [mock_w3]),
-            call(self.cls, 'svc2', mock_limit4, [mock_c2], []),
+            call(
+                'svc1', mock_limit1, [mock_c1], [mock_w1],
+                colorize=False
+            ),
+            call(
+                'svc1', mock_limit2, [], [mock_w2], colorize=False
+            ),
+            call(
+                'svc2', mock_limit3, [], [mock_w3], colorize=False
+            ),
+            call(
+                'svc2', mock_limit4, [mock_c2], [], colorize=False
+            ),
         ]
         assert mock_d2c.mock_calls == [
             call({
@@ -931,12 +941,12 @@ class TestCheckThresholds(RunnerTester):
         }
         mock_checker.get_limits.return_value = {}
 
-        def se_print(cls, s, l, c, w):
+        def se_print(s, l, c, w, colorize=True):
             return ('{s}/{l}'.format(s=s, l=l.name), '')
 
         self.cls.checker = mock_checker
         self.cls.skip_check = ['svc1/limit1']
-        with patch('awslimitchecker.runner.Runner.print_issue',
+        with patch('%s.issue_string_tuple' % pb,
                    autospec=True) as mock_print:
             mock_print.side_effect = se_print
             with patch('awslimitchecker.runner.dict2cols') as mock_d2c:
@@ -947,7 +957,9 @@ class TestCheckThresholds(RunnerTester):
             call.check_thresholds(use_ta=True, service=None)
         ]
         assert mock_print.mock_calls == [
-            call(self.cls, 'svc1', mock_limit2, [], [mock_w2]),
+            call(
+                'svc1', mock_limit2, [], [mock_w2], colorize=True
+            ),
         ]
         assert mock_d2c.mock_calls == [
             call({
@@ -986,7 +998,7 @@ class TestCheckThresholds(RunnerTester):
         mock_checker.get_limits.return_value = {}
 
         self.cls.checker = mock_checker
-        with patch('awslimitchecker.runner.Runner.print_issue',
+        with patch('%s.issue_string_tuple' % pb,
                    autospec=True) as mock_print:
             mock_print.return_value = ('', '')
             with patch('awslimitchecker.runner.dict2cols') as mock_d2c:
@@ -996,8 +1008,13 @@ class TestCheckThresholds(RunnerTester):
             call.check_thresholds(use_ta=True, service=None)
         ]
         assert mock_print.mock_calls == [
-            call(self.cls, 'svc1', mock_limit1, [], [mock_w1, mock_w2]),
-            call(self.cls, 'svc2', mock_limit2, [], [mock_w3]),
+            call(
+                'svc1', mock_limit1, [], [mock_w1, mock_w2],
+                colorize=True
+            ),
+            call(
+                'svc2', mock_limit2, [], [mock_w3], colorize=True
+            ),
         ]
         assert res == (1, {
             'svc2': {
@@ -1031,7 +1048,7 @@ class TestCheckThresholds(RunnerTester):
 
         self.cls.checker = mock_checker
         self.cls.service_name = ['svc2']
-        with patch('awslimitchecker.runner.Runner.print_issue',
+        with patch('%s.issue_string_tuple' % pb,
                    autospec=True) as mock_print:
             mock_print.return_value = ('', '')
             with patch('awslimitchecker.runner.dict2cols') as mock_d2c:
@@ -1041,7 +1058,9 @@ class TestCheckThresholds(RunnerTester):
             call.check_thresholds(use_ta=True, service=['svc2'])
         ]
         assert mock_print.mock_calls == [
-            call(self.cls, 'svc2', mock_limit2, [], [mock_w3]),
+            call(
+                'svc2', mock_limit2, [], [mock_w3], colorize=True
+            ),
         ]
         assert res == (1, {
             'svc2': {
@@ -1067,7 +1086,7 @@ class TestCheckThresholds(RunnerTester):
 
         self.cls.checker = mock_checker
         self.cls.skip_ta = True
-        with patch('awslimitchecker.runner.Runner.print_issue',
+        with patch('%s.issue_string_tuple' % pb,
                    autospec=True) as mock_print:
             mock_print.return_value = ('', '')
             with patch('awslimitchecker.runner.dict2cols') as mock_d2c:
@@ -1077,173 +1096,16 @@ class TestCheckThresholds(RunnerTester):
             call.check_thresholds(use_ta=False, service=None)
         ]
         assert mock_print.mock_calls == [
-            call(self.cls, 'svc1', mock_limit1, [mock_c1, mock_c2], []),
+            call(
+                'svc1', mock_limit1, [mock_c1, mock_c2], [],
+                colorize=True
+            ),
         ]
         assert res == (2, {
             'svc1': {
                 'limit1': mock_limit1,
             },
         }, '  \n')
-
-
-class TestPrintIssue(RunnerTester):
-
-    def test_crit_one(self):
-        mock_limit = Mock(spec_set=AwsLimit)
-        type(mock_limit).name = 'limitname'
-        mock_limit.get_limit.return_value = 12
-
-        c1 = AwsLimitUsage(mock_limit, 56)
-
-        def se_color(s, c, colorize=True):
-            return 'xX%sXx' % s
-
-        with patch('%s.color_output' % pb) as m_co:
-            m_co.side_effect = se_color
-            res = self.cls.print_issue(
-                'svcname',
-                mock_limit,
-                [c1],
-                []
-            )
-        assert res == ('svcname/limitname',
-                       '(limit 12) xXCRITICAL: 56Xx')
-        assert m_co.mock_calls == [
-            call('CRITICAL: 56', 'red', colorize=True)
-        ]
-
-    def test_crit_multi(self):
-        mock_limit = Mock(spec_set=AwsLimit)
-        type(mock_limit).name = 'limitname'
-        mock_limit.get_limit.return_value = 5
-
-        c1 = AwsLimitUsage(mock_limit, 10)
-        c2 = AwsLimitUsage(mock_limit, 12, resource_id='c2id')
-        c3 = AwsLimitUsage(mock_limit, 8)
-
-        def se_color(s, c, colorize=True):
-            return 'xX%sXx' % s
-
-        with patch('%s.color_output' % pb) as m_co:
-            m_co.side_effect = se_color
-            res = self.cls.print_issue(
-                'svcname',
-                mock_limit,
-                [c1, c2, c3],
-                []
-            )
-        assert res == ('svcname/limitname',
-                       '(limit 5) xXCRITICAL: 8, 10, c2id=12Xx')
-        assert m_co.mock_calls == [
-            call('CRITICAL: 8, 10, c2id=12', 'red', colorize=True)
-        ]
-
-    def test_warn_one(self):
-        mock_limit = Mock(spec_set=AwsLimit)
-        type(mock_limit).name = 'limitname'
-        mock_limit.get_limit.return_value = 12
-
-        w1 = AwsLimitUsage(mock_limit, 11)
-
-        def se_color(s, c, colorize=True):
-            return 'xX%sXx' % s
-
-        with patch('%s.color_output' % pb) as m_co:
-            m_co.side_effect = se_color
-            res = self.cls.print_issue(
-                'svcname',
-                mock_limit,
-                [],
-                [w1]
-            )
-        assert res == ('svcname/limitname', '(limit 12) xXWARNING: 11Xx')
-        assert m_co.mock_calls == [
-            call('WARNING: 11', 'yellow', colorize=True)
-        ]
-
-    def test_warn_multi(self):
-        mock_limit = Mock(spec_set=AwsLimit)
-        type(mock_limit).name = 'limitname'
-        mock_limit.get_limit.return_value = 12
-
-        w1 = AwsLimitUsage(mock_limit, 11)
-        w2 = AwsLimitUsage(mock_limit, 10, resource_id='w2id')
-        w3 = AwsLimitUsage(mock_limit, 10, resource_id='w3id')
-
-        def se_color(s, c, colorize=True):
-            return 'xX%sXx' % s
-
-        with patch('%s.color_output' % pb) as m_co:
-            m_co.side_effect = se_color
-            res = self.cls.print_issue(
-                'svcname',
-                mock_limit,
-                [],
-                [w1, w2, w3]
-            )
-        assert res == ('svcname/limitname',
-                       '(limit 12) xXWARNING: w2id=10, w3id=10, 11Xx')
-        assert m_co.mock_calls == [
-            call('WARNING: w2id=10, w3id=10, 11', 'yellow', colorize=True)
-        ]
-
-    def test_both_one(self):
-        self.cls.colorize = False
-        mock_limit = Mock(spec_set=AwsLimit)
-        type(mock_limit).name = 'limitname'
-        mock_limit.get_limit.return_value = 12
-
-        c1 = AwsLimitUsage(mock_limit, 10)
-        w1 = AwsLimitUsage(mock_limit, 10, resource_id='w3id')
-
-        def se_color(s, c, colorize=True):
-            return 'xX%sXx' % s
-
-        with patch('%s.color_output' % pb) as m_co:
-            m_co.side_effect = se_color
-            res = self.cls.print_issue(
-                'svcname',
-                mock_limit,
-                [c1],
-                [w1]
-            )
-        assert res == ('svcname/limitname',
-                       '(limit 12) xXCRITICAL: 10Xx xXWARNING: w3id=10Xx')
-        assert m_co.mock_calls == [
-            call('CRITICAL: 10', 'red', colorize=False),
-            call('WARNING: w3id=10', 'yellow', colorize=False)
-        ]
-
-    def test_both_multi(self):
-        mock_limit = Mock(spec_set=AwsLimit)
-        type(mock_limit).name = 'limitname'
-        mock_limit.get_limit.return_value = 12
-
-        c1 = AwsLimitUsage(mock_limit, 10)
-        c2 = AwsLimitUsage(mock_limit, 12, resource_id='c2id')
-        c3 = AwsLimitUsage(mock_limit, 8)
-        w1 = AwsLimitUsage(mock_limit, 11)
-        w2 = AwsLimitUsage(mock_limit, 10, resource_id='w2id')
-        w3 = AwsLimitUsage(mock_limit, 10, resource_id='w3id')
-
-        def se_color(s, c, colorize=True):
-            return 'xX%sXx' % s
-
-        with patch('%s.color_output' % pb) as m_co:
-            m_co.side_effect = se_color
-            res = self.cls.print_issue(
-                'svcname',
-                mock_limit,
-                [c1, c2, c3],
-                [w1, w2, w3]
-            )
-        assert res == ('svcname/limitname',
-                       '(limit 12) xXCRITICAL: 8, 10, c2id=12Xx '
-                       'xXWARNING: w2id=10, w3id=10, 11Xx')
-        assert m_co.mock_calls == [
-            call('CRITICAL: 8, 10, c2id=12', 'red', colorize=True),
-            call('WARNING: w2id=10, w3id=10, 11', 'yellow', colorize=True)
-        ]
 
 
 class TestConsoleEntryPoint(RunnerTester):
@@ -2059,7 +1921,9 @@ class TestConsoleEntryPoint(RunnerTester):
             with patch('%s.Runner.check_thresholds' % pb,
                        autospec=True) as mock_ct:
                 mock_ct.return_value = 0, {}, 'foo'
-                with patch('%s.color_output' % pb) as m_co:
+                with patch(
+                    'awslimitchecker.utils.color_output'
+                ) as m_co:
                     m_co.return_value = 'COLORIZED'
                     with pytest.raises(SystemExit):
                         self.cls.console_entry_point()
